myFunctionApp/requirements.txt
azure-functions
textblob
openai
python-dotenv